file clean
