+++ playground/sft/qwen3/qwen3_1p7b_sft_step3_data.py
@@ -24,9 +24,9 @@ class Exp(BaseExp):
         self.scheduler_cfg.warmup_schedule = 100
 
         self.checkpoint_cfg.load_option.none(but=["model"])
-        self.checkpoint_cfg.load_safetensors = "/mnt/step2-alignment-jfs/zane/opensources_model/Qwen3-1.7B-Base/"
+        self.checkpoint_cfg.load_safetensors = "/oss/opensources_model/Qwen3-1.7B-Base/"
         self.checkpoint_cfg.save_safetensors = True
-        self.checkpoint_cfg.save_dir = "/mnt/shared-storage/tenant/tmp/zhy/tmp/"
+        self.checkpoint_cfg.save_dir = "/oss/checkpoints/"
         self.checkpoint_cfg.save_option.all()
         self.checkpoint_cfg.save_interval = 1000
