+++ playground/sft/step3/step3_toy_sft_step3_data.py
@@ -20,9 +20,9 @@ class Exp(BaseExp):
         self.trainer_cfg.log_interval = 1
 
         self.checkpoint_cfg.load_option.none(but=["model"])
-        # self.checkpoint_cfg.load_safetensors = "/mnt/step2-alignment-jfs/zane/opensources_model/Qwen3-30B-A3B-Base"
+        # self.checkpoint_cfg.load_safetensors = "/oss/opensources_model/Qwen3-30B-A3B-Base"
         self.checkpoint_cfg.save_safetensors = True
-        self.checkpoint_cfg.save_dir = "/mnt/shared-storage/tenant/tmp/zhy/tmp/"
+        self.checkpoint_cfg.save_dir = "/oss/checkpoints/"
         self.checkpoint_cfg.save_option.all()
         self.checkpoint_cfg.save_interval = 100
         self.model_cfg.recompute = True
